+++ app/src/main/kotlin/com/looker/kenko/data/local/dao/SessionDao.kt
@@ -51,11 +51,11 @@ interface SessionDao {
 
     @Query(
         """
-        SELECT COUNT(*)
+        SELECT MIN(date)
         FROM sessions
         """,
     )
-    suspend fun getTotalSessions(): Int
+    suspend fun earliestSessionDate(): EpochDays?
 
     @Query(
         """