+++ app/src/main/kotlin/com/looker/kenko/data/repository/PlanRepo.kt
@@ -63,8 +63,6 @@ interface PlanRepo {
 
     suspend fun deletePlan(id: Int)
 
-    suspend fun deleteEmptyPlans()
-
     suspend fun addItem(planItem: PlanItem)
 
     suspend fun removeItem(id: Long)